Navigation

+++ app/src/main/java/me/tbandawa/android/openweather/service/Coordinates.kt
@@ -1,6 +1,6 @@
 package me.tbandawa.android.openweather.service
 
 data class Coordinates(
-    var latitude: Float,
-    var longitude: Float
+    var latitude: Double,
+    var longitude: Double
 )
