+++ domain/src/main/java/openweather/domain/repository/OpenWeatherRepository.kt
@@ -7,7 +7,7 @@ interface OpenWeatherRepository {
 
     suspend fun fetchCurrentWeather(city: String): Flow<NetworkResult<CurrentWeather>>
 
-    suspend fun fetchOneCall(lat: Long, lon: Long): Flow<NetworkResult<OneCall>>
+    suspend fun fetchOneCall(lat: Double, lon: Double): Flow<NetworkResult<OneCall>>
 
     suspend fun fetchFiveDayWeather(city: String): Flow<NetworkResult<FiveDayWeatherForecast>>
